minor bugs solved

+++ acstracks_app/models.py
@@ -120,12 +120,8 @@ class Preference(models.Model):
     default_profile = models.CharField(
         max_length=255, null=True, blank=True, default=None
         )
-    maximum_heart_rate = models.DecimalField(
-        max_digits=3, decimal_places=0, default=175
-        )
-    resting_heart_rate = models.DecimalField(
-        max_digits=3, decimal_places=0, default=70
-        )
+    maximum_heart_rate = models.IntegerField(default=175)
+    resting_heart_rate = models.IntegerField(default=70)
 
     
     def __str__(self):
